Allow batch editing of tags.
